more renaming from localdocker to localpodman
